Mapbox test (#36)
* Origin and Destination

* Origin, Destination complete. Missing POST

* POST implemented

* Better back handling

* Fixed reference to HttpHandler that was renamed on master

* Newest SHA1 keys

+++ app/src/main/java/cit/jauc/lib/CoordsConverter.java
@@ -4,7 +4,7 @@ import cit.jauc.model.Location;
 
 public class CoordsConverter {
 
-    public static String getLocationfromCoords(long longitude, long latitude) {
+    public static String getLocationfromCoords(double longitude, double latitude) {
         // TODO API Call to get location from coords
         return "Location " + longitude + "," + latitude;
     }
